+++ DawnLib.Dusk/src/API/Definitions/Dungeon/DuskDungeonDefinition.cs
@@ -70,8 +70,8 @@ public class DuskDungeonDefinition : DuskContentDefinition<DawnDungeonInfo>
             builder.SetAssetBundlePath(mod.GetRelativePath("Assets", DungeonFlowReference.BundleName));
             builder.SetMapTileSize(MapTileSize);
             builder.SetFirstTimeAudio(StingerAudio);
-            builder.SetStingerPlaysMoreThanOnce(StingerPlaysMoreThanOnce);
-            builder.SetStingerPlayChance(StingerPlayChance);
+            builder.OverrideStingerPlaysMoreThanOnce(StingerPlaysMoreThanOnce);
+            builder.OverrideStingerPlayChance(StingerPlayChance);
             builder.SetDungeonRangeClamp(DungeonRangeClamp);
             builder.SetWeights(weightBuilder => weightBuilder.SetGlobalWeight(SpawnWeights));
             ApplyTagsTo(builder);